Automation fix for GCE UI based on feature changes
