+++ tests/foreman/ui/test_host.py
@@ -19,7 +19,6 @@
 import copy
 import csv
 import os
-import random
 import re
 from datetime import datetime
 
@@ -1860,282 +1859,6 @@ def test_positive_delete_libvirt(
         assert session.host.search(name)[0]['Name'] == name
         session.host.delete(name)
         assert not target_sat.api.Host().search(query={'search': f'name="{hostname}"'})
-
-
-@pytest.fixture
-def gce_template(googleclient):
-    max_rhel7_template = max(
-        img.name for img in googleclient.list_templates(True) if str(img.name).startswith('rhel-7')
-    )
-    return googleclient.get_template(max_rhel7_template, project='rhel-cloud').uuid
-
-
-@pytest.fixture
-def gce_cloudinit_template(googleclient, gce_cert):
-    return googleclient.get_template('customcinit', project=gce_cert['project_id']).uuid
-
-
-@pytest.fixture
-def gce_domain(module_org, smart_proxy_location, gce_cert, target_sat):
-    domain_name = f'{settings.gce.zone}.c.{gce_cert["project_id"]}.internal'
-    domain = target_sat.api.Domain().search(query={'search': f'name={domain_name}'})
-    if domain:
-        domain = domain[0]
-        domain.organization = [module_org]
-        domain.location = [smart_proxy_location]
-        domain.update(['organization', 'location'])
-    if not domain:
-        domain = target_sat.api.Domain(
-            name=domain_name, location=[smart_proxy_location], organization=[module_org]
-        ).create()
-    return domain
-
-
-@pytest.fixture
-def gce_resource_with_image(
-    gce_template,
-    gce_cloudinit_template,
-    gce_cert,
-    default_architecture,
-    default_os,
-    smart_proxy_location,
-    module_org,
-    target_sat,
-):
-    with Session('gce_tests') as session:
-        # Until the CLI and API support is added for GCE,
-        # creating GCE CR from UI
-        cr_name = gen_string('alpha')
-        vm_user = gen_string('alpha')
-        session.computeresource.create(
-            {
-                'name': cr_name,
-                'provider': FOREMAN_PROVIDERS['google'],
-                'provider_content.google_project_id': gce_cert['project_id'],
-                'provider_content.client_email': gce_cert['client_email'],
-                'provider_content.certificate_path': settings.gce.cert_path,
-                'provider_content.zone.value': settings.gce.zone,
-                'organizations.resources.assigned': [module_org.name],
-                'locations.resources.assigned': [smart_proxy_location.name],
-            }
-        )
-    gce_cr = target_sat.api.AbstractComputeResource().search(query={'search': f'name={cr_name}'})[0]
-    # Finish Image
-    target_sat.api.Image(
-        architecture=default_architecture,
-        compute_resource=gce_cr,
-        name='autogce_img',
-        operatingsystem=default_os,
-        username=vm_user,
-        uuid=gce_template,
-    ).create()
-    # Cloud-Init Image
-    target_sat.api.Image(
-        architecture=default_architecture,
-        compute_resource=gce_cr,
-        name='autogce_img_cinit',
-        operatingsystem=default_os,
-        username=vm_user,
-        uuid=gce_cloudinit_template,
-        user_data=True,
-    ).create()
-    return gce_cr
-
-
-@pytest.fixture
-def gce_hostgroup(
-    module_org,
-    smart_proxy_location,
-    default_partition_table,
-    default_architecture,
-    default_os,
-    gce_domain,
-    gce_resource_with_image,
-    module_lce,
-    module_cv_repo,
-    target_sat,
-):
-    return target_sat.api.HostGroup(
-        architecture=default_architecture,
-        compute_resource=gce_resource_with_image,
-        domain=gce_domain,
-        lifecycle_environment=module_lce,
-        content_view=module_cv_repo,
-        location=[smart_proxy_location],
-        operatingsystem=default_os,
-        organization=[module_org],
-        ptable=default_partition_table,
-    ).create()
-
-
-@pytest.mark.tier4
-@pytest.mark.run_in_one_thread
-@pytest.mark.skip_if_not_set('gce')
-@pytest.mark.parametrize('setting_update', ['destroy_vm_on_host_delete'], indirect=True)
-def test_positive_gce_provision_end_to_end(
-    session,
-    target_sat,
-    module_org,
-    smart_proxy_location,
-    default_os,
-    gce_domain,
-    gce_hostgroup,
-    googleclient,
-    setting_update,
-    remove_vm_on_delete,
-):
-    """Provision Host on GCE compute resource
-
-    :id: 8d1877bb-fbc2-4969-a13e-e95e4df4f4cd
-
-    :expectedresults: Host is provisioned successfully
-
-    :CaseLevel: System
-    """
-    name = f'test{gen_string("alpha", 4).lower()}'
-    hostname = f'{name}.{gce_domain.name}'
-    gceapi_vmname = hostname.replace('.', '-')
-    root_pwd = gen_string('alpha', 15)
-    storage = [{'size': 20}]
-    with Session('gce_tests') as session:
-        session.organization.select(org_name=module_org.name)
-        session.location.select(loc_name=smart_proxy_location.name)
-        # Provision GCE Host
-        try:
-            with target_sat.skip_yum_update_during_provisioning(
-                template='Kickstart default finish'
-            ):
-                session.host.create(
-                    {
-                        'host.name': name,
-                        'host.hostgroup': gce_hostgroup.name,
-                        'provider_content.virtual_machine.machine_type': 'g1-small',
-                        'provider_content.virtual_machine.external_ip': True,
-                        'provider_content.virtual_machine.network': 'default',
-                        'provider_content.virtual_machine.storage': storage,
-                        'operating_system.operating_system': default_os.title,
-                        'operating_system.image': 'autogce_img',
-                        'operating_system.root_password': root_pwd,
-                    }
-                )
-                wait_for(
-                    lambda: target_sat.api.Host()
-                    .search(query={'search': f'name={hostname}'})[0]
-                    .build_status_label
-                    != 'Pending installation',
-                    timeout=600,
-                    delay=15,
-                    silent_failure=True,
-                    handle_exception=True,
-                )
-                # 1. Host Creation Assertions
-                # 1.1 UI based Assertions
-                host_info = session.host.get_details(hostname)
-                assert session.host.search(hostname)[0]['Name'] == hostname
-                assert host_info['properties']['properties_table']['Build'] == 'Installed clear'
-                # 1.2 GCE Backend Assertions
-                gceapi_vm = googleclient.get_vm(gceapi_vmname)
-                assert gceapi_vm.is_running
-                assert gceapi_vm
-                assert gceapi_vm.name == gceapi_vmname
-                assert gceapi_vm.zone == settings.gce.zone
-                assert gceapi_vm.ip == host_info['properties']['properties_table']['IP Address']
-                assert 'g1-small' in gceapi_vm.raw['machineType'].split('/')[-1]
-                assert 'default' in gceapi_vm.raw['networkInterfaces'][0]['network'].split('/')[-1]
-                # 2. Host Deletion Assertions
-                session.host.delete(hostname)
-                assert not target_sat.api.Host().search(query={'search': f'name="{hostname}"'})
-                # 2.2 GCE Backend Assertions
-                assert gceapi_vm.is_stopping or gceapi_vm.is_stopped
-        except Exception as error:
-            gcehost = target_sat.api.Host().search(query={'search': f'name={hostname}'})
-            if gcehost:
-                gcehost[0].delete()
-            raise error
-        finally:
-            googleclient.disconnect()
-
-
-@pytest.mark.tier4
-@pytest.mark.upgrade
-@pytest.mark.run_in_one_thread
-@pytest.mark.skip_if_not_set('gce')
-@pytest.mark.parametrize('setting_update', ['destroy_vm_on_host_delete'], indirect=True)
-def test_positive_gce_cloudinit_provision_end_to_end(
-    session,
-    target_sat,
-    module_org,
-    smart_proxy_location,
-    default_os,
-    gce_domain,
-    gce_hostgroup,
-    googleclient,
-    setting_update,
-    remove_vm_on_delete,
-):
-    """Provision Host on GCE compute resource
-
-    :id: 6ee63ec6-2e8e-4ed6-ae48-e68b078233c6
-
-    :expectedresults: Host is provisioned successfully
-
-    :CaseLevel: System
-    """
-    name = f'test{gen_string("alpha", 4).lower()}'
-    hostname = f'{name}.{gce_domain.name}'
-    gceapi_vmname = hostname.replace('.', '-')
-    storage = [{'size': 20}]
-    root_pwd = gen_string('alpha', random.choice([8, 15]))
-    with Session('gce_tests') as session:
-        session.organization.select(org_name=module_org.name)
-        session.location.select(loc_name=smart_proxy_location.name)
-        # Provision GCE Host
-        try:
-            with target_sat.skip_yum_update_during_provisioning(
-                template='Kickstart default user data'
-            ):
-                session.host.create(
-                    {
-                        'host.name': name,
-                        'host.hostgroup': gce_hostgroup.name,
-                        'provider_content.virtual_machine.machine_type': 'g1-small',
-                        'provider_content.virtual_machine.external_ip': True,
-                        'provider_content.virtual_machine.network': 'default',
-                        'provider_content.virtual_machine.storage': storage,
-                        'operating_system.operating_system': default_os.title,
-                        'operating_system.image': 'autogce_img_cinit',
-                        'operating_system.root_password': root_pwd,
-                    }
-                )
-                # 1. Host Creation Assertions
-                # 1.1 UI based Assertions
-                host_info = session.host.get_details(hostname)
-                assert session.host.search(hostname)[0]['Name'] == hostname
-                assert (
-                    host_info['properties']['properties_table']['Build']
-                    == 'Pending installation clear'
-                )
-                # 1.2 GCE Backend Assertions
-                gceapi_vm = googleclient.get_vm(gceapi_vmname)
-                assert gceapi_vm
-                assert gceapi_vm.is_running
-                assert gceapi_vm.name == gceapi_vmname
-                assert gceapi_vm.zone == settings.gce.zone
-                assert gceapi_vm.ip == host_info['properties']['properties_table']['IP Address']
-                assert 'g1-small' in gceapi_vm.raw['machineType'].split('/')[-1]
-                assert 'default' in gceapi_vm.raw['networkInterfaces'][0]['network'].split('/')[-1]
-                # 2. Host Deletion Assertions
-                session.host.delete(hostname)
-                assert not target_sat.api.Host().search(query={'search': f'name="{hostname}"'})
-                # 2.2 GCE Backend Assertions
-                assert gceapi_vm.is_stopping or gceapi_vm.is_stopped
-        except Exception as error:
-            gcehost = target_sat.api.Host().search(query={'search': f'name={hostname}'})
-            if gcehost:
-                gcehost[0].delete()
-            raise error
-        finally:
-            googleclient.disconnect()
 
 
 # ------------------------------ NEW HOST UI DETAILS ----------------------------
